+++ src/main/java/nextstep/path/application/PathService.java
@@ -4,6 +4,9 @@ import nextstep.line.application.LineProvider;
 import nextstep.line.domain.Line;
 import nextstep.path.application.dto.PathResponse;
 import nextstep.path.application.dto.PathSearchRequest;
+import nextstep.path.application.fare.extra.distance.FareChain;
+import nextstep.path.application.fare.extra.distance.FirstExtraFareHandler;
+import nextstep.path.application.fare.extra.distance.SecondExtraFareHandler;
 import nextstep.path.domain.Path;
 import nextstep.path.domain.SubwayMap;
 import nextstep.path.exception.PathNotFoundException;
