+++ src/main/java/nextstep/path/application/fare/extra/distance/FareChain.java
@@ -1,4 +1,4 @@
-package nextstep.path.application;
+package nextstep.path.application.fare.extra.distance;
 
 import nextstep.path.exception.FareApplyingException;
