+++ src/main/java/nextstep/path/application/fare/extra/distance/SecondExtraFareHandler.java
@@ -1,6 +1,6 @@
-package nextstep.path.application;
+package nextstep.path.application.fare.extra.distance;
 
-public class FirstExtraFareHandler extends PathFareHandler {
+public class SecondExtraFareHandler extends PathFareHandler {
     private static final long EXTRA_FARE_BASE = 100L;
 
     @Override
@@ -21,12 +21,11 @@ public class FirstExtraFareHandler extends PathFareHandler {
 
     @Override
     protected int getStandardDistance() {
-        return 10;
+        return 50;
     }
 
     @Override
     protected int getFareInterval() {
-        return 5;
+        return 8;
     }
-
 }
